Add tslint and fix ALL the formatting
Wait---we are doing some Rust-related thing *and* we're using a fancy
tool to get a type system *but* weren't adding ALL the lints? Gotta fix
that!

Also, let me take a second to write down my appreciation for TSLint's
defaults---it has trailing commas. And if you are struggling with it,
just add `--fix`.